Update 060_forget.rst
Added overview description of difference between keep and keep-within. Also information that those not marked as keep will be deleted.

It was not clear to me from the documentation, that it functions like a marking algorithm marking stuff-to-keep by different rules and then deletes the rest. I hope it is more clear with the short description I added.

Also, I was very confused of the different between keep-* and keep-within-*, hence the overview at the top.